:recycle: SectionInputReply 컴포넌트 분리
